Create README.md
We got the concept of e-commerce from the shops that are available for online shops. Online business is gaining a lot of popularity these days and we see that there are various product posts on Facebook in different groups / pages which are of interest to the general public. But we have caught a lot of problems in our eyes on these issues. Different pages / groups on Facebook do not have these products in a well-arranged manner, due to which we have to face many problems. Besides, our concept is for the problems like product information is not given